requirements.txt
flask==2.3.3
openai==1.3.0
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
langchain==0.0.350
langchain-openai==0.0.2
streamlit==1.28.1
pandas==2.1.1
numpy==1.24.3
requests==2.31.0
beautifulsoup4==4.12.2
markdown==3.5.1
